pyproject.toml
[project]
name = "monograph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.12",
    "sqlalchemy>=2.0.41",
    "strawberry-graphql[fastapi]>=0.269.0",
    "vi-core",
]

[tool.uv.sources]
vi-core = { git = "https://git.vashinvestor.ru/razvitie/microservices/vi-core.git" }

[tool.ruff]
line-length = 120

[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
implicit_reexport = true
warn_return_any = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
